lg-apigateway-oc/requirements.txt
fastapi
httpx
python-jose[cryptography]
python-dotenv
uvicorn
